==> studies/study2_N40/6_study2_N40_results.qmd <==
---
title: "Analyzing bayesNMF results on study 2 data with a maximum of N = 40 signatures"
author: "Jenna Landy"
---

```{r}
library(tidyverse)
library(ggplot2)
library(bayesNMF)
library(ggh4x)
library(cowplot)
library(glue)

data_dir <- "../../data/study2"
output_dir <- "../../output/study2_N40"
processed_dir <- "../../processed/study2_N40"
figures_dir <- "../../figures/study2_N40"
```

```{r}
results <- read.csv(file.path(processed_dir, "metrics.csv")) %>%
  filter(
    grepl('signatureanalyzer', model) | grepl('SBFI', model) | grepl('BIC', model)
  )

results %>%
  group_by(model) %>%
  tally()
```

## Results 

```{r}
rank_metrics_aligned = results %>%
  mutate(
    model = factor(model, levels = names(color_list))
  ) %>%
  mutate(
    rank_bias = learned_rank - N,
  ) %>%
  pivot_longer(c(precision, sensitivity, rank_bias, total_minutes), names_to = "metric", values_to = "value") %>%
  mutate(
    metric = factor(metric, levels = c('rank_bias','precision','sensitivity', 'total_minutes'), labels = c('Rank Bias','Precision','Sensitivity', 'Minutes'))
  ) %>%
  ggplot(aes(x = as.factor(N), y = value, color = model)) +
  facet_grid(rows = vars(metric), scales = 'free', switch = 'y') +
  geom_boxplot() +
  labs(
    x = "True Rank",
    color = "Model"
  ) +
  theme(
    text = element_text(size = 20),
    legend.position = "top",
    legend.justification = "left",
    axis.title.y = element_blank(),
    strip.text.y.left = element_text(angle = 0),
    strip.placement = 'outside'
  ) +
  scale_color_manual(
    breaks = names(color_list),
    labels = color_names,
    values = color_list
  ) +
  guides(
    color = guide_legend(nrow = 2, byrow = TRUE)
  ) +
  theme_light(base_size = 12) +
  theme(
    legend.position = "none",
    axis.title.y = element_blank(),
    strip.background = element_rect(fill = "white", color = "black", linewidth = 0.5),
    strip.text = element_text(color = 'black'),
    plot.background = element_rect(fill = "white", color = "black", linewidth = 0.5),
    plot.title.position = "plot"
  ) +
  guides(
    color = guide_legend(nrow = 2, byrow = TRUE)
  ) +
  ggtitle("A. Learning rank in 1-40")
  ggh4x::facetted_pos_scales(
    y = list(NULL, scale_y_continuous(limits = c(0, 1.05)), scale_y_continuous(limits = c(0, 1.05)), scale_y_continuous(trans = "log2"))
  )
saveRDS(rank_metrics_aligned, file.path(figures_dir, "N40_rank_metrics_aligned.rds"))
ggsave(file.path(figures_dir, "N40_rank_metrics_aligned.png"), rank_metrics_aligned, width = 7, height = 7)
```
